move era and player to new folder
