--- src/components/Era.ts
@@ -1,8 +1,0 @@
-export enum Era {
-    SEVENTIES = "1968-1979",
-    EIGHTIES = "1980-1989",
-    NINETIES = "1990-1999",
-    TWO_THOUSAND = "2000-2009",
-    TWO_THOUSAND_TENS = "2010-Pres.",
-    ALL = "All Years",
-}
