+++ src/components/board/Board.tsx
@@ -1,4 +1,4 @@
-import { Player } from "../Player";
+import { Player } from "../../models/Player";
 import "./Board.css";
 
 const POSITIONS = {
