+++ src/components/pop-ups/game-over/GameOverPopUp.tsx
@@ -1,4 +1,4 @@
-import { Player } from "../../Player";
+import { Player } from "../../../models/Player";
 import "./GameOverPopUp.css";
 
 interface GameOverPopUpProps {
